Refactor : 변수명 수정
